+++ featuristic/synthesis/__init__.py
@@ -5,4 +5,4 @@ from .fitness import fitness_pearson
 from .mrmr import MaxRelevanceMinRedundancy
 from .engine import deserialize_program, evaluate_programs, run_genetic_algorithm
 from .render import render_prog
-from .symbolic_functions import CustomSymbolicFunction, list_symbolic_functions
+from .symbolic_functions import list_symbolic_functions